[breadboard-cli] Extract the debug-server to make it easier to maintain
